+++ CHANGELOG.md
@@ -65,6 +65,10 @@ Happy April Fool's Day! While we did genuinely rework the settings page structur
 # LoLLMs v19.0 (alpha) Changelog
 *Date: March 06, 2025*
 
+## [2026-03-15 14:34]
+
+- fix(google_imagen): move aspect_ratio into config object
+
 ## [2026-03-15 14:30]
 
 - chore: remove functions_zoo and services_zoo submodule pull logic